initial project structure refactor
